moved go code into src/
