+++ docgen-2/src/components/nav.templ
@@ -1,7 +1,7 @@
 package components
 
 import "strings"
-import r "github.com/a-h/templ/docgen-2/render"
+import r "github.com/a-h/templ/docgen-2/src/render"
 
 templ nav(c *r.PageContext, pages []*r.Page, depth int) {
 	<nav
