+++ docgen-2/src/components/structure.templ
@@ -1,6 +1,6 @@
 package components
 
-import r "github.com/a-h/templ/docgen-2/render"
+import r "github.com/a-h/templ/docgen-2/src/render"
 
 templ HTML(c *r.PageContext, pages []*r.Page, data string) {
 	<html>
