+++ docgen-2/src/render/section.go
