+++ docgen-2/src/render/type.go
